services/gateway/requirements.txt
# api we choose as the gateway
fastapi==0.115.0
uvicorn[standard]==0.30.6
# template from FastAPI
jinja2==3.1.4
# grpc tools
grpcio==1.76.0
grpcio-tools==1.76.0
python-multipart==0.0.9
protobuf>=6.31.1,<7.0.0
